pyproject.toml
[project]
name = "tool-authorization-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-arcade>=1.3.1",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.0",
    "python-dotenv>=1.0.0",
]
